A bag of small (but useful) fixes (#202)
* Do not unnecessarily pass around eos_id

The end-of-sequence marker is now handled by a stopping condition, so
it's not used by the generator anymore.

* Do not assume inference in Generator

Remove the `torch.no_grad` wrapping and switching the model to
eval mode. The generator could be used for training in the future,
so we should not impose an inference regime.

Add `eval` and `train` methods to all generator classes.

* Fixup AutoGenerator test

* Attention mask dtype must match query/value dtype

This avoids accidental widening plus Torch SDP requires it.

* Revert "Do not assume inference in Generator"

This reverts commit c7129148b2a2bb563ed45cf4ae1b9d6384059537.

We want to hash out the training API more and see if we can perhaps
separate inference/training in such a way that the user does not
have to change the model mode or run without gradient tracking.

+++ curated_transformers/models/attention.py
@@ -83,7 +83,7 @@ class AttentionMask:
 
     def logit_mask(self, dtype: torch.dtype):
         if self._logit_mask is None:
-            self._logit_mask = (1.0 - self.bool_mask.int()) * torch.finfo(dtype).min
+            self._logit_mask = (1.0 - self.bool_mask.to(dtype)) * torch.finfo(dtype).min
         return self._logit_mask
 
     @property
